registry/requirements.txt
beautifulsoup4==4.10.0
html5lib==1.1
mysql-connector-python==8.1.0
python-decouple==3.7
